pyproject.toml
[project]
name = "flame"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "backoff>=2.2.1",
    "datasets>=3.5.1",
    "dvc>=3.59.2",
    "evaluate>=0.4.3",
    "huggingface-hub>=0.30.2",
    "litellm==1.67.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pyyaml>=6.0.2",
    "ruff>=0.11.8",
    "scikit-learn>=1.6.1",
    "setuptools>=80.1.0",
    "tenacity>=9.1.2",
    "tiktoken>=0.9.0",
    "together>=1.5.5",
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "matplotlib>=3.10.1",
    "openai>=1.76.2",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "requests-mock>=1.12.1",
    "ruff>=0.11.8",
    "seaborn>=0.13.2",
    "bert-score>=0.3.13",
]

[tool.pytest.ini_options]
filterwarnings = [
  "ignore::pydantic._internal._config.PydanticDeprecatedSince20",
  "ignore::DeprecationWarning",
  "ignore::sklearn.exceptions.UndefinedMetricWarning",
]

[tool.ruff]
exclude = ["**/*.ipynb"]

[dependency-groups]
dev = [
    "huggingface-hub>=0.30.2",
    "ruff>=0.11.8",
]
